+++ pkg/clientset/typed/konnect/v1alpha1/fake/fake_konnect_client.go
@@ -36,6 +36,10 @@ func (c *FakeKonnectV1alpha1) KonnectCloudGatewayNetworks(namespace string) v1al
 	return newFakeKonnectCloudGatewayNetworks(c, namespace)
 }
 
+func (c *FakeKonnectV1alpha1) KonnectExtensions(namespace string) v1alpha1.KonnectExtensionInterface {
+	return newFakeKonnectExtensions(c, namespace)
+}
+
 func (c *FakeKonnectV1alpha1) KonnectGatewayControlPlanes(namespace string) v1alpha1.KonnectGatewayControlPlaneInterface {
 	return newFakeKonnectGatewayControlPlanes(c, namespace)
 }