+++ pkg/clientset/typed/konnect/v1alpha1/generated_expansion.go
@@ -22,4 +22,6 @@ type KonnectAPIAuthConfigurationExpansion interface{}
 
 type KonnectCloudGatewayNetworkExpansion interface{}
 
+type KonnectExtensionExpansion interface{}
+
 type KonnectGatewayControlPlaneExpansion interface{}